Add test for operator/pkg/util RandomString function
Add tests for RandomString function with three different lengths of a
negative number, zero and positive number.
The function panics when a negative number is passed as an argument.

A check is added to util.RandomString function to check if the argument
is less than zero and return empty string if the argument is a negative
number